# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Badminton Service API
  version: 0.0.1
tags:
  - name: TelegramAuth
paths:
  /api/auth/telegram/login:
    post:
      summary: Login via Telegram authorization data
      operationId: telegramLogin
      tags:
        - TelegramAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramAuthData'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TelegramAuthData'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramUser'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/telegram/logout:
    post:
      summary: Logout
      operationId: telegramLogout
      tags:
        - TelegramAuth
      responses:
        '204':
          description: Logged out
  /api/auth/telegram/user:
    get:
      summary: Get current user from cookie
      operationId: telegramCurrentUser
      tags:
        - TelegramAuth
      parameters:
        - name: tg_user
          in: cookie
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramUser'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    TelegramAuthData:
      type: object
      required:
        - id
        - auth_date
        - hash
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
        photo_url:
          type: string
        auth_date:
          type: integer
          format: int64
        hash:
          type: string
    TelegramUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
        photo_url:
          type: string
        auth_date:
          type: integer
          format: int64
        hash:
          type: string
    ErrorResponse:
      type: object
      required:
        - timestamp
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        path:
          type: string